feat: move specs to engine and add optimizer alias
