+++ src/main/java/org/example/flowmod/AppLauncher.java
@@ -2,8 +2,8 @@ package org.example.flowmod;
 
 import com.google.gson.Gson;
 import org.example.flowmod.engine.FlowPhysics;
-import org.example.flowmod.engine.GraduatedHoleOptimizer;
-import org.example.flowmod.engine.DesignResult;
+
+import org.example.flowmod.HoleOptimizer;
 import org.example.flowmod.HoleLayout;
 import org.example.flowmod.HoleSpec;
 
@@ -21,15 +21,13 @@ public class AppLauncher {
         }
 
         FlowPhysics.Result phys = FlowPhysics.compute(settings.pipe());
-        GraduatedHoleOptimizer opt = new GraduatedHoleOptimizer();
-        DesignResult result = opt.optimise(
+        HoleLayout layout = HoleOptimizer.optimise(
                 settings.pipe(),
                 settings.filter(),
                 settings.drillMinMm(),
                 settings.drillMaxMm(),
                 settings.rows(),
                 settings.Cd());
-        HoleLayout layout = result.holeLayout();
 
         System.out.println("Row | Pos mm | \u00D8 mm | L/min");
         for (HoleSpec h : layout.holes()) {
