--- src/main/java/org/example/flowmod/FilterSpecs.java
@@ -1,3 +1,0 @@
-package org.example.flowmod;
-
-public record FilterSpecs(double openAreaPercent, double faceVelocityMaxMs) {}
