+++ src/main/java/org/example/flowmod/FlowModifierUI.java
@@ -9,6 +9,8 @@ import javafx.scene.layout.VBox;
 import javafx.scene.paint.Color;
 import javafx.stage.Stage;
 import org.example.flowmod.engine.FlowPhysics;
+import org.example.flowmod.engine.FilterSpecs;
+import org.example.flowmod.engine.PipeSpecs;
 
 import java.util.ArrayList;
 import java.util.List;
