+++ src/main/java/org/example/flowmod/Settings.java
@@ -1,5 +1,8 @@
 package org.example.flowmod;
 
+import org.example.flowmod.engine.FilterSpecs;
+import org.example.flowmod.engine.PipeSpecs;
+
 public record Settings(
         PipeSpecs pipe,
         FilterSpecs filter,
