+++ src/main/java/org/example/flowmod/engine/FilterSpecs.java
@@ -1,0 +1,4 @@
+package org.example.flowmod.engine;
+
+/** Specifications for the outlet filter. */
+public record FilterSpecs(double openAreaPercent, double faceVelocityMaxMs) {}
