+++ src/main/java/org/example/flowmod/engine/ModifierDesignStrategy.java
@@ -1,7 +1,7 @@
 package org.example.flowmod.engine;
 
-import org.example.flowmod.FilterSpecs;
-import org.example.flowmod.PipeSpecs;
+import org.example.flowmod.engine.FilterSpecs;
+import org.example.flowmod.engine.PipeSpecs;
 
 public interface ModifierDesignStrategy {
     DesignResult optimise(PipeSpecs pipe,
