+++ src/main/java/org/example/flowmod/engine/PipeSpecs.java
@@ -1,3 +1,4 @@
-package org.example.flowmod;
+package org.example.flowmod.engine;
 
+/** Specifications for the upstream pipe and modifier. */
 public record PipeSpecs(double innerDiameterMm, double flowRateLpm, double modifierLengthMm) {}
