+++ src/main/java/org/example/flowmod/engine/VaneDiffuserOptimizer.java
@@ -1,7 +1,7 @@
 package org.example.flowmod.engine;
 
-import org.example.flowmod.FilterSpecs;
-import org.example.flowmod.PipeSpecs;
+import org.example.flowmod.engine.FilterSpecs;
+import org.example.flowmod.engine.PipeSpecs;
 
 import java.util.ArrayList;
 import java.util.List;
