+++ src/main/java/org/example/flowmod/model3d/ModelBuilder.java
@@ -2,7 +2,7 @@ package org.example.flowmod.model3d;
 
 import javafx.scene.shape.MeshView;
 import javafx.scene.shape.TriangleMesh;
-import org.example.flowmod.PipeSpecs;
+import org.example.flowmod.engine.PipeSpecs;
 import org.example.flowmod.engine.DesignResult;
 
 public final class ModelBuilder {
