+++ src/test/java/org/example/flowmod/FlowPhysicsTest.java
@@ -3,6 +3,7 @@ package org.example.flowmod;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
 import org.example.flowmod.engine.FlowPhysics;
+import org.example.flowmod.engine.PipeSpecs;
 
 public class FlowPhysicsTest {
     @Test
